+++ home-manager/default.nix
@@ -3,5 +3,6 @@
     ../common/keyboard.nix
     ./modules/identity.nix
     ./modules/programs
+    ./modules/services
   ];
 }